[SYCL] Remove old scheduler
