+++ sycl/include/CL/sycl/accessor.hpp
@@ -1,4 +1,4 @@
-//==--------- accessor.hpp --- SYCL accessor -------------------------------==//
+//==------------ accessor.hpp - SYCL standard header file ------------------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,1032 +6,827 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifdef SCHEDULER_20
-
-#include <CL/sycl/accessor2.hpp>
-
-#else
-
 #pragma once
-
-#include <type_traits>
 
 #include <CL/sycl/atomic.hpp>
 #include <CL/sycl/buffer.hpp>
+#include <CL/sycl/detail/accessor_impl.hpp>
 #include <CL/sycl/detail/common.hpp>
 #include <CL/sycl/handler.hpp>
 #include <CL/sycl/id.hpp>
 #include <CL/sycl/pointers.hpp>
-#include <CL/sycl/detail/scheduler/scheduler.h>
-#include <CL/sycl/stl.hpp>
+
+// The file contains implementations of accessor class. Objects of accessor
+// class define a requirement to access some SYCL memory object or local memory
+// of the device.
+//
+// Basically there are 3 distinct types of accessors.
+//
+// One of them is an accessor to a SYCL buffer object(Buffer accessor) which has
+// the richest interface. It supports things like accessing only a part of
+// buffer, multidimensional access using sycl::id, conversions to various
+// multi_ptr and atomic classes.
+//
+// Second type is an accessor to a SYCL image object(Image accessor) which has
+// "image" specific methods for reading and writing.
+//
+// Finally, accessor to local memory(Local accessor) doesn't require access to
+// any SYCL memory object, but asks for some local memory on device to be
+// available. Some methods overlap with ones that "Buffer accessor" provides.
+//
+// Buffer and Image accessors create the requirement to access some SYCL memory
+// object(or part of it). SYCL RT must detect when two kernels want to access
+// the same memory objects and make sure they are executed in correct order.
+//
+// "accessor_common" class that contains several common methods between Buffer
+// and Local accessors.
+//
+// Accessors have different representation on host and on device. On host they
+// have non-templated base class, that is needed to safely work with any
+// accessor type. Furhermore on host we need some additional fields in order
+// to implement functionality required by Specification, for example during
+// lifetime of a host accessor other operations with memory object the accessor
+// refers to should be blocked and when all references to the host accessor are
+// desctructed, the memory this host accessor refers to should be "written
+// back".
+//
+// The scheme of inheritance for host side:
+//
+//  +------------------+     +-----------------+     +-----------------------+
+//  |                  |     |                 |     |                       |
+//  | AccessorBaseHost |     | accessor_common |     | LocalAccessorBaseHost |
+//  |                  |     |                 |     |                       |
+//  +------------------+     +-----+-----------+     +--------+--------------+
+//         |     |                     |   |                   |
+//         |     +-----------+    +----+   +---------+  +------+
+//         |                 |    |                  |  |
+//         v                 v    v                  v  v
+//  +----------------+  +-----------------+   +-------------+
+//  |                |  |   accessor(1)   |   | accessor(3) |
+//  | image_accessor |  +-----------------|   +-------------+
+//  |                |  | for targets:    |   | for target: |
+//  +---+---+---+----+  |                 |   |             |
+//      |   |   |       | host_buffer     |   | local       |
+//      |   |   |       | global_buffer   |   +-------------+
+//      |   |   |       | constant_buffer |
+//      |   |   |       +-----------------+
+//      |   |   |
+//      |   |   +------------------------------------+
+//      |   |                                        |
+//      |   +----------------------+                 |
+//      v                          v                 v
+//  +-----------------+    +--------------+    +-------------+
+//  |     acessor(2)  |    |  accessor(4) |    | accessor(5) |
+//  +-----------------+    +--------------+    +-------------+
+//  | for targets:    |    | for targets: |    | for target: |
+//  |                 |    |              |    |             |
+//  | host_image      |    |  image       |    | image_array |
+//  +-----------------+    +--------------+    +-------------+
+//
+// For host side AccessorBaseHost/LocalAccessorBaseHost contains shared_ptr
+// which points to AccessorImplHost/LocalAccessorImplHost object.
+//
+// The scheme of inheritance for device side:
+//
+//                            +-----------------+
+//                            |                 |
+//                            | accessor_common |
+//                            |                 |
+//                            +-----+-------+---+
+//                                     |       |
+//                                +----+       +-----+
+//                                |                  |
+//                                v                  v
+//  +----------------+  +-----------------+   +-------------+
+//  |                |  |   accessor(1)   |   | accessor(3) |
+//  | image_accessor |  +-----------------|   +-------------+
+//  |                |  | for targets:    |   | for target: |
+//  +---+---+---+----+  |                 |   |             |
+//      |   |   |       | host_buffer     |   | local       |
+//      |   |   |       | global_buffer   |   +-------------+
+//      |   |   |       | constant_buffer |
+//      |   |   |       +-----------------+
+//      |   |   |
+//      |   |   +------------------------------------+
+//      |   |                                        |
+//      |   +----------------------+                 |
+//      v                          v                 v
+//  +-----------------+    +--------------+    +-------------+
+//  |     acessor(2)  |    |  accessor(4) |    | accessor(5) |
+//  +-----------------+    +--------------+    +-------------+
+//  | for targets:    |    | for targets: |    | for target: |
+//  |                 |    |              |    |             |
+//  | host_image      |    |  image       |    | image_array |
+//  +-----------------+    +--------------+    +-------------+
+//
+// For device side AccessorImplHost/LocalAccessorImplHost are fileds of
+// accessor(1) and accessor(3).
+//
+// accessor(1) declares accessor as a template class and implements accessor
+// class for access targets: host_buffer, global_buffer and constant_buffer.
+//
+// accessor(3) specializes accessor(1) for the local access target.
+//
+// image_accessor contains implements interfaces for access targets: host_image,
+// image and image_array. But there are three distinct specializations of the
+// accessor(1) (accessor(2), accessor(4), accessor(5)) that are just inherited
+// from image_accessor.
+//
+// accessor_common contains several helpers common for both accessor(1) and
+// accessor(3)
 
 namespace cl {
 namespace sycl {
+
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::target AccessTarget = access::target::global_buffer,
+          access::placeholder IsPlaceholder = access::placeholder::false_t>
+class accessor;
+
 namespace detail {
 
-template <typename dataT, int dimensions, access::mode accessMode,
-          access::target accessTarget, access::placeholder isPlaceholder>
-class accessor_base;
+// The function extends or truncates number of dimensions of objects of id
+// or ranges classes. When extending the new values are filled with
+// DefaultValue, truncation just removes extra values.
+template <int NewDim, int DefaultValue, template <int> class T, int OldDim>
+static T<NewDim> convertToArrayOfN(T<OldDim> OldObj) {
+  T<NewDim> NewObj;
+  const int CopyDims = NewDim > OldDim ? OldDim : NewDim;
+  for (int I = 0; I < CopyDims; ++I)
+    NewObj[I] = OldObj[I];
+  for (int I = CopyDims; I < NewDim; ++I)
+    NewObj[I] = DefaultValue;
+  return NewObj;
+}
 
-template <int accessorDim, typename dataT, int dimensions,
-          access::mode accessMode, access::target accessTarget,
-          access::placeholder isPlaceholder>
-class subscript_obj {
-  using accessor_t = accessor_base<dataT, accessorDim, accessMode, accessTarget,
-                                   isPlaceholder>;
-
-  // TODO: Remove reference here as subscript_obj, can potentially outlive
-  // the accessor. There is no spec-defined usecase, so leave it for now.
-  const accessor_t &accRef;
-  cl::sycl::id<accessorDim> ids;
-
-public:
-  subscript_obj(const accessor_t &acc, cl::sycl::id<accessorDim> &indexes)
-      : accRef(acc), ids(indexes) {}
-
-  subscript_obj<accessorDim, dataT, dimensions - 1, accessMode, accessTarget,
-                isPlaceholder>
-  operator[](size_t index) {
-    ids[accessorDim - dimensions] = index;
-    return subscript_obj<accessorDim, dataT, dimensions - 1, accessMode,
-                         accessTarget, isPlaceholder>(accRef, ids);
-  }
-};
-
-template <int accessorDim, typename dataT, access::mode accessMode,
-          access::target accessTarget, access::placeholder isPlaceholder>
-class subscript_obj<accessorDim, dataT, 1, accessMode, accessTarget,
-                    isPlaceholder> {
-  using accessor_t = accessor_base<dataT, accessorDim, accessMode, accessTarget,
-                                   isPlaceholder>;
-
-  const accessor_t &accRef;
-  cl::sycl::id<accessorDim> ids;
-
-public:
-  subscript_obj(const accessor_t &acc, cl::sycl::id<accessorDim> &indexes)
-      : accRef(acc), ids(indexes) {}
-
-  dataT &operator[](size_t index) {
-    ids[accessorDim - 1] = index;
-    return accRef.__get_impl()->Data[getOffsetForId(
-      accRef.__get_impl()->MemRange, ids, accRef.__get_impl()->Offset)];
-  }
-};
-
-template <int accessorDim, typename dataT,
-          access::target accessTarget, access::placeholder isPlaceholder>
-class subscript_obj<accessorDim, dataT, 1, access::mode::read, accessTarget,
-                    isPlaceholder> {
-  using accessor_t = accessor_base<dataT, accessorDim, access::mode::read,
-                                   accessTarget, isPlaceholder>;
-
-  const accessor_t &accRef;
-  cl::sycl::id<accessorDim> ids;
-
-public:
-  subscript_obj(const accessor_t &acc, cl::sycl::id<accessorDim> &indexes)
-      : accRef(acc), ids(indexes) {}
-
-  typename detail::remove_AS<dataT>::type
-  operator[](size_t index) {
-    ids[accessorDim - 1] = index;
-    return accRef.__get_impl()->Data[getOffsetForId(
-      accRef.__get_impl()->MemRange, ids, accRef.__get_impl()->Offset)];
-  }
-};
-
-/// Specializations of accessor_impl define data fields for accessor.
-/// There is no default implementation for the class. This class is
-/// not a root of the class hierarchy, because it should be
-/// initialized at the bottom of the hierarchy.
-template <typename dataT, int dimensions, access::mode accessMode,
-          access::target accessTarget, access::placeholder isPlaceholder,
-          typename voidT = void>
-struct accessor_impl;
-
-#define SYCL_ACCESSOR_IMPL(CONDITION)                                          \
-  template <typename dataT, int dimensions, access::mode accessMode,           \
-            access::target accessTarget, access::placeholder isPlaceholder>    \
-  struct accessor_impl<dataT, dimensions, accessMode, accessTarget,            \
-                       isPlaceholder,                                          \
-                       typename std::enable_if<(CONDITION)>::type>
-
-/// Implementation of host accessor providing access to a single element.
-/// Available when (dimensions == 0).
-SYCL_ACCESSOR_IMPL(isTargetHostAccess(accessTarget) && dimensions == 0) {
-  dataT *Data;
-  // For simplicity of the padding in the main accessor class
-  // we ensure that accessor_impl is not smaller than size of shared_ptr.
-  char padding[sizeof(std::shared_ptr<accessor_impl>) - sizeof(dataT *)];
-  accessor_impl(dataT *Data) : Data(Data) {}
-
-  // Returns the number of accessed elements.
-  size_t get_count() const { return 1; }
-
-#ifdef __SYCL_DEVICE_ONLY__
-  bool operator==(const accessor_impl &Rhs) const { return Data == Rhs.Data; }
-#endif
-};
-
-/// Implementation of host accessor.
-/// Available when (dimensions > 0).
-SYCL_ACCESSOR_IMPL(isTargetHostAccess(accessTarget) && dimensions > 0) {
-  dataT *Data;
-  // Accessor's own range, can be subset of MemRange
-  range<dimensions> AccessRange;
-  // Range of corresponding memory object
-  range<dimensions> MemRange;
-  id<dimensions> Offset;
-
-  accessor_impl(dataT * Data, range<dimensions> AccessRange,
-                range<dimensions> MemRange, id<dimensions> Offset = {})
-      : Data(Data), AccessRange(AccessRange), MemRange(MemRange),
-        Offset(Offset) {}
-
-  // Returns the number of accessed elements.
-  size_t get_count() const { return AccessRange.size(); }
-
-#ifdef __SYCL_DEVICE_ONLY__
-  bool operator==(const accessor_impl &Rhs) const {
-    return Data == Rhs.Data && AccessRange == Rhs.AccessRange &&
-           MemRange == Rhs.MemRange && Offset == Rhs.Offset;
-  }
-#endif
-};
-
-/// Implementation of device (kernel) accessor providing access to a single
-/// element. Available only when (dimensions == 0).
-/// There is no way to tell at compile time if this accessor will be used
-/// on OpenCL device or on host. So, the class should fit both variants.
-SYCL_ACCESSOR_IMPL(!isTargetHostAccess(accessTarget) &&
-                   accessTarget != access::target::local &&
-                   dimensions == 0) {
-  // This field must be the first to guarantee that it's safe to use
-  // reinterpret casting while setting kernel arguments in order to get cl_mem
-  // value from the buffer regardless of the accessor's dimensionality.
-#ifndef __SYCL_DEVICE_ONLY__
-  detail::buffer_impl<buffer_allocator> *m_Buf = nullptr;
-#else
-  char padding[sizeof(detail::buffer_impl<buffer_allocator> *)];
-#endif // __SYCL_DEVICE_ONLY__
-
-  dataT *Data;
-
-  // Device accessor constructor.
-  accessor_impl(dataT *Data) : Data(Data) {}
-
-  // Returns the number of accessed elements.
-  size_t get_count() const { return 1; }
-
-#ifdef __SYCL_DEVICE_ONLY__
-  bool operator==(const accessor_impl &Rhs) const { return Data == Rhs.Data; }
-#endif
-
-  static_assert(
-      std::is_same<typename DeviceValueType<dataT, accessTarget>::type,
-                   dataT>::value,
-      "The type should have been adjusted before propagating through "
-      "class hierarchy");
-};
-
-/// Implementation of device (kernel) accessor. There is no way to
-/// tell at compile time if this accessor will be used on OpenCL
-/// device or on host. So, the class should fit both variants.
-/// Available only when (dimensions > 0).
-SYCL_ACCESSOR_IMPL(!isTargetHostAccess(accessTarget) &&
-                   accessTarget != access::target::local &&
-                   dimensions > 0) {
-  // This field must be the first to guarantee that it's safe to use
-  // reinterpret casting while setting kernel arguments in order to get cl_mem
-  // value from the buffer regardless of the accessor's dimensionality.
-#ifndef __SYCL_DEVICE_ONLY__
-  detail::buffer_impl<buffer_allocator> *m_Buf = nullptr;
-#else
-  char padding[sizeof(detail::buffer_impl<buffer_allocator> *)];
-#endif // __SYCL_DEVICE_ONLY__
-
-  dataT *Data;
-  // Accessor's own range, can be subset of MemRange
-  range<dimensions> AccessRange;
-  // Range of corresponding memory object
-  range<dimensions> MemRange;
-  id<dimensions> Offset;
-
-  // Device accessor constructor.
-  accessor_impl(dataT * Data, range<dimensions> AccessRange,
-                range<dimensions> MemRange, id<dimensions> Offset = {})
-      : Data(Data), AccessRange(AccessRange), MemRange(MemRange),
-        Offset(Offset) {}
-
-  // Returns the number of accessed elements.
-  size_t get_count() const { return AccessRange.size(); }
-
-#ifdef __SYCL_DEVICE_ONLY__
-  bool operator==(const accessor_impl &Rhs) const {
-    return Data == Rhs.Data && AccessRange == Rhs.AccessRange &&
-           MemRange == Rhs.MemRange && Offset == Rhs.Offset;
-  }
-#endif
-
-  static_assert(
-      std::is_same<typename DeviceValueType<dataT, accessTarget>::type,
-                   dataT>::value,
-      "The type should have been adjusted before propagating through "
-      "class hierarchy");
-};
-
-/// Implementation of local accessor providing access to a single element.
-/// Available only when (dimensions == 0).
-SYCL_ACCESSOR_IMPL(accessTarget == access::target::local &&
-                   dimensions == 0) {
-  // This field must be the first to guarantee that it's safe to use
-  // reinterpret casting while setting kernel arguments in order to get size
-  // value from the accessor regardless of its dimensionality.
-  size_t ByteSize;
-
-#ifndef __SYCL_DEVICE_ONLY__
-  shared_ptr_class<vector_class<dataT>> dataBuf;
-#else
-  char padding[sizeof(shared_ptr_class<vector_class<dataT>>)];
-#endif
-
-  dataT *Data;
-
-  accessor_impl(handler &Handler)
-      : ByteSize(sizeof(dataT)) {
-#ifndef __SYCL_DEVICE_ONLY__
-    if (Handler.is_host()) {
-      dataBuf = std::make_shared<vector_class<dataT>>(1);
-      Data = dataBuf->data();
-    }
-#endif
-  }
-
-  // Returns the number of accessed elements.
-  size_t get_count() const { return 1; }
-
-#ifdef __SYCL_DEVICE_ONLY__
-  bool operator==(const accessor_impl &Rhs) const {
-    return ByteSize == Rhs.ByteSize && Data == Rhs.Data;
-  }
-#endif
-
-  static_assert(
-      std::is_same<typename DeviceValueType<dataT, accessTarget>::type,
-                   dataT>::value,
-      "The type should have been adjusted before propagating through "
-      "class hierarchy");
-};
-
-/// Implementation of local accessor.
-/// Available only when (dimensions > 0).
-SYCL_ACCESSOR_IMPL(accessTarget == access::target::local &&
-                   dimensions > 0) {
-  // This field must be the first to guarantee that it's safe to use
-  // reinterpret casting while setting kernel arguments in order to get size
-  // value from the accessor regardless of its dimensionality.
-  size_t ByteSize;
-
-#ifndef __SYCL_DEVICE_ONLY__
-  shared_ptr_class<vector_class<dataT>> dataBuf;
-#else
-  char padding[sizeof(shared_ptr_class<vector_class<dataT>>)];
-#endif
-
-  dataT *Data;
-  // Accessor's own range
-  range<dimensions> AccessRange;
-  // For local accessor AccessRange and MemRange always are same but both fields
-  // are presented here to keep no differences between local and global
-  // accessors for compiler
-  range<dimensions> MemRange;
-  // TODO delete it when accessor class was remade
-  // Offset field is not need for local accessor, but this field is now used
-  // in the inheritance hierarchy. Getting rid of this field will cause
-  // duplication and complication of the code even more.
-  id<dimensions> Offset;
-
-  accessor_impl(range<dimensions> Range, handler &Handler)
-      : AccessRange(Range), MemRange(Range),
-        ByteSize(Range.size() * sizeof(dataT)) {
-#ifndef __SYCL_DEVICE_ONLY__
-    if (Handler.is_host()) {
-      dataBuf = std::make_shared<vector_class<dataT>>(Range.size());
-      Data = dataBuf->data();
-    }
-#endif
-  }
-
-  // Returns the number of accessed elements.
-  size_t get_count() const { return AccessRange.size(); }
-
-#ifdef __SYCL_DEVICE_ONLY__
-  bool operator==(const accessor_impl &Rhs) const {
-    return ByteSize == Rhs.ByteSize && Data == Rhs.Data &&
-           AccessRange == Rhs.AccessRange && MemRange == Rhs.MemRange &&
-           Offset == Rhs.Offset;
-  }
-#endif
-
-  static_assert(
-      std::is_same<typename DeviceValueType<dataT, accessTarget>::type,
-                   dataT>::value,
-      "The type should have been adjusted before propagating through "
-      "class hierarchy");
-};
-
-/// Base class for all accessor specializations.
-template <typename dataT, int dimensions, access::mode accessMode,
-          access::target accessTarget, access::placeholder isPlaceholder>
-class accessor_base {
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::target AccessTarget, access::placeholder IsPlaceholder>
+class accessor_common {
 protected:
-  template <int, typename, int, access::mode, access::target,
-            access::placeholder>
-  friend class subscript_obj;
-  friend class ::cl::sycl::simple_scheduler::Node;
-  friend class ::cl::sycl::simple_scheduler::Scheduler;
-  using _ImplT =
-      accessor_impl<dataT, dimensions, accessMode, accessTarget, isPlaceholder>;
+  constexpr static bool IsPlaceH = IsPlaceholder == access::placeholder::true_t;
+  constexpr static access::address_space AS = TargetToAS<AccessTarget>::AS;
 
-  const _ImplT *__get_impl() const {
-#ifdef __SYCL_DEVICE_ONLY__
-    return reinterpret_cast<const _ImplT *>(this);
-#else
-    auto ImplPtrPtr = reinterpret_cast<const std::shared_ptr<_ImplT> *>(this);
-    const _ImplT* I = &**ImplPtrPtr;
-    return I;
-#endif
-  }
+  constexpr static bool IsHostBuf = AccessTarget == access::target::host_buffer;
 
-  _ImplT *__get_impl() {
-#ifdef __SYCL_DEVICE_ONLY__
-    return reinterpret_cast<_ImplT *>(this);
-#else
-    auto ImplPtrPtr = reinterpret_cast<std::shared_ptr<_ImplT> *>(this);
-    _ImplT* I = &**ImplPtrPtr;
-    return I;
-#endif
-  }
+  constexpr static bool IsGlobalBuf =
+      AccessTarget == access::target::global_buffer;
 
-  static_assert(
-      std::is_same<typename DeviceValueType<dataT, accessTarget>::type,
-                   dataT>::value,
-      "The type should have been adjusted before propagating through "
-      "class hierarchy");
-};
+  constexpr static bool IsConstantBuf =
+      AccessTarget == access::target::constant_buffer;
 
-// The macro is used to conditionally define methods of accessor class
-// by wrapping them into a structure that is non-empty only if the
-// condition is met.
-#define SYCL_ACCESSOR_SUBCLASS(TAG, PARENT, CONDITION)                         \
-  template <typename dataT, int dimensions, access::mode accessMode,           \
-            access::target accessTarget, access::placeholder isPlaceholder,    \
-            typename voidT = void>                                             \
-  struct TAG : ::cl::sycl::detail::PARENT<dataT, dimensions, accessMode,       \
-                                          accessTarget, isPlaceholder> {};     \
-                                                                               \
-  template <typename dataT, int dimensions, access::mode accessMode,           \
-            access::target accessTarget, access::placeholder isPlaceholder>    \
-  struct TAG<dataT, dimensions, accessMode, accessTarget, isPlaceholder,       \
-             typename std::enable_if<(CONDITION)>::type>                       \
-      : ::cl::sycl::detail::PARENT<dataT, dimensions, accessMode,              \
-                                   accessTarget, isPlaceholder>
+  constexpr static bool IsAccessAnyWrite =
+      AccessMode == access::mode::write ||
+      AccessMode == access::mode::read_write ||
+      AccessMode == access::mode::discard_write ||
+      AccessMode == access::mode::discard_read_write;
 
-SYCL_ACCESSOR_SUBCLASS(accessor_common, accessor_base, true /* always */) {
-  // Returns true if the current accessor is a placeholder accessor.
-  constexpr bool is_placeholder() const {
-    return isPlaceholder == access::placeholder::true_t;
-  }
+  constexpr static bool IsAccessReadOnly = AccessMode == access::mode::read;
 
-  // Returns the size of the accessed memory in bytes.
-  size_t get_size() const { return this->get_count() * sizeof(dataT); }
+  constexpr static bool IsAccessReadWrite =
+      AccessMode == access::mode::read_write;
 
-  // Returns the number of accessed elements.
-  size_t get_count() const { return this->__get_impl()->get_count(); }
+  using RefType = typename detail::PtrValueType<DataT, AS>::type &;
+  using PtrType = typename detail::PtrValueType<DataT, AS>::type *;
 
-  template <int Dimensions = dimensions>
-  typename std::enable_if<(Dimensions > 0), range<Dimensions>>::type
-  get_range() const { return this->__get_impl()->AccessRange; }
+  using AccType =
+      accessor<DataT, Dimensions, AccessMode, AccessTarget, IsPlaceholder>;
 
-  template <int Dimensions = dimensions>
-  typename std::enable_if<(Dimensions > 0), id<Dimensions>>::type
-  get_offset() const { return this->__get_impl()->Offset; }
-};
+  // The class which allows to access value of N dimensional accessor using N
+  // subscript operators, e.g. accessor[2][2][3]
+  template <int SubDims> class AccessorSubscript {
+    static constexpr int Dims = Dimensions;
 
-SYCL_ACCESSOR_SUBCLASS(accessor_opdata_w, accessor_common,
-                       (accessMode == access::mode::write ||
-                        accessMode == access::mode::read_write ||
-                        accessMode == access::mode::discard_write ||
-                        accessMode == access::mode::discard_read_write) &&
-                       dimensions == 0) {
-  operator dataT &() const {
-    return this->__get_impl()->Data[0];
-  }
-};
+    template <bool B, class T = void>
+    using enable_if_t = typename std::enable_if<B, T>::type;
 
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_wn, accessor_opdata_w,
-                       (accessMode == access::mode::write ||
-                        accessMode == access::mode::read_write ||
-                        accessMode == access::mode::discard_write ||
-                        accessMode == access::mode::discard_read_write) &&
-                       dimensions > 0) {
-  dataT &operator[](id<dimensions> index) const {
-    return this->__get_impl()->Data[getOffsetForId(
-      this->__get_impl()->MemRange, index, this->get_offset())];
-  }
+    mutable id<Dims> MIDs;
+    AccType MAccessor;
 
-  subscript_obj<dimensions, dataT, dimensions - 1, accessMode, accessTarget,
-              isPlaceholder>
-  operator[](size_t index) const {
-    id<dimensions> ids;
-    ids[0] = index;
-    return subscript_obj<dimensions, dataT, dimensions - 1, accessMode,
-                         accessTarget, isPlaceholder>(*this, ids);
-  }
-};
+  public:
+    AccessorSubscript(AccType Accessor, id<Dims> IDs)
+        : MAccessor(Accessor), MIDs(IDs) {}
 
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_w, accessor_subscript_wn,
-                       (accessMode == access::mode::write ||
-                        accessMode == access::mode::read_write ||
-                        accessMode == access::mode::discard_write ||
-                        accessMode == access::mode::discard_read_write) &&
-                       dimensions == 1) {
-  // The tricky part here is that there is no function overloading
-  // between different scopes in C++. That is, operator[] defined in a
-  // child class hides any operator[] defined in any of the parent
-  // classes. That's why operator[] defined in accessor_subscript_wn
-  // is not visible here and we have to define
-  // operator[](id<dimensions>) once again.
-  dataT &operator[](id<dimensions> index) const {
-    return this->operator[](
-      getOffsetForId(this->__get_impl()->MemRange, index, this->get_offset()));
-  }
-  dataT &operator[](size_t index) const {
-    return this->__get_impl()->Data[index];
-  }
-};
+    // Only accessor class is supposed to use this c'tor for the first
+    // operator[].
+    AccessorSubscript(AccType Accessor, size_t Index) : MAccessor(Accessor) {
+      MIDs[0] = Index;
+    }
 
-SYCL_ACCESSOR_SUBCLASS(accessor_opdata_r, accessor_subscript_w,
-                       accessMode == access::mode::read && dimensions == 0) {
-  using PureType = typename detail::remove_AS<dataT>::type;
-  operator PureType() const {
-    return this->__get_impl()->Data[0];
-  }
-};
+    template <int CurDims = SubDims, typename = enable_if_t<(CurDims > 1)>>
+    AccessorSubscript<CurDims - 1> operator[](size_t Index) {
+      MIDs[Dims - CurDims] = Index;
+      return AccessorSubscript<CurDims - 1>(MAccessor, MIDs);
+    }
 
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_rn, accessor_opdata_r,
-                       accessMode == access::mode::read && dimensions > 0) {
-  typename detail::remove_AS<dataT>::type
-  operator[](id<dimensions> index) const {
-    return this->__get_impl()->Data[getOffsetForId(
-      this->__get_impl()->MemRange, index, this->get_offset())];
-  }
+    template <int CurDims = SubDims,
+              typename = enable_if_t<IsAccessAnyWrite && CurDims == 1>>
+    RefType operator[](size_t Index) const {
+      MIDs[Dims - CurDims] = Index;
+      return MAccessor[MIDs];
+    }
 
-  subscript_obj<dimensions, dataT, dimensions - 1, accessMode, accessTarget,
-              isPlaceholder>
-  operator[](size_t index) const {
-    id<dimensions> ids;
-    ids[0] = index;
-    return subscript_obj<dimensions, dataT, dimensions - 1, accessMode,
-                         accessTarget, isPlaceholder>(*this, ids);
-  }
-};
-
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_r, accessor_subscript_rn,
-                       accessMode == access::mode::read && dimensions == 1) {
-  typename detail::remove_AS<dataT>::type
-  operator[](id<dimensions> index) const {
-    return this->operator[](
-      getOffsetForId(this->__get_impl()->MemRange, index, this->get_offset()));
-  }
-  typename detail::remove_AS<dataT>::type
-  operator[](size_t index) const {
-    return this->__get_impl()->Data[index];
-  }
-};
-
-template <access::target accessTarget> struct getAddressSpace {
-  constexpr static cl::sycl::access::address_space value =
-      cl::sycl::access::address_space::global_space;
-};
-
-template <> struct getAddressSpace<access::target::local> {
-  constexpr static cl::sycl::access::address_space value =
-      cl::sycl::access::address_space::local_space;
-};
-
-// Available when: accessMode == access::mode::atomic && dimensions == 0
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_atomic_eq0, accessor_subscript_r,
-                       accessMode == access::mode::atomic && dimensions == 0) {
-  using PureType = typename detail::remove_AS<dataT>::type;
-  constexpr static access::address_space addressSpace =
-      getAddressSpace<accessTarget>::value;
-  operator atomic<PureType, addressSpace>() const {
-    return atomic<PureType, addressSpace>(
-        multi_ptr<PureType, addressSpace>(&(this->__get_impl()->Data[0])));
-  }
-};
-
-// Available when: accessMode == access::mode::atomic && dimensions > 0
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_atomic_gt0,
-                       accessor_subscript_atomic_eq0,
-                       accessMode == access::mode::atomic && dimensions > 0) {
-  using PureType = typename detail::remove_AS<dataT>::type;
-  constexpr static access::address_space addressSpace =
-      getAddressSpace<accessTarget>::value;
-  atomic<PureType, addressSpace> operator[](id<dimensions> index) const {
-    return atomic<PureType, addressSpace>(
-        multi_ptr<PureType, addressSpace>(&(this->__get_impl()->Data[getOffsetForId(
-            this->__get_impl()->MemRange, index, this->__get_impl()->Offset)])));
-  }
-};
-
-// Available only when: accessMode == access::mode::atomic && dimensions == 1
-SYCL_ACCESSOR_SUBCLASS(accessor_subscript_atomic_eq1,
-                       accessor_subscript_atomic_gt0,
-                       accessMode == access::mode::atomic && dimensions == 1) {
-  using PureType = typename detail::remove_AS<dataT>::type;
-  constexpr static access::address_space addressSpace =
-      getAddressSpace<accessTarget>::value;
-  atomic<PureType, addressSpace> operator[](size_t index) const {
-    return atomic<PureType, addressSpace>(
-        multi_ptr<PureType, addressSpace>(&(this->__get_impl()->Data[index])));
-  }
-};
-
-// TODO:
-// /* Available only when: dimensions > 1 */
-// __unspecified__ &operator[](size_t index) const;
-
-SYCL_ACCESSOR_SUBCLASS(accessor_pointer, accessor_subscript_atomic_eq1, true) {
-  /* Available only when: accessTarget == access::target::host_buffer */
-  template <typename DataT = typename detail::remove_AS<dataT>::type,
-            access::target AccessTarget = accessTarget>
-  typename std::enable_if<(AccessTarget == access::target::host_buffer),
-                          dataT *>::type
-  get_pointer() const {
-    return this->__get_impl()->Data;
-  }
-  /* Available only when: accessTarget == access::target::global_buffer */
-  template <typename DataT = typename detail::remove_AS<dataT>::type,
-            access::target AccessTarget = accessTarget>
-  typename std::enable_if<(AccessTarget == access::target::global_buffer),
-                          global_ptr<DataT>>::type
-  get_pointer() const {
-    return global_ptr<DataT>(this->__get_impl()->Data);
-  }
-  /* Available only when: accessTarget == access::target::constant_buffer */
-  template <typename DataT = typename detail::remove_AS<dataT>::type,
-            access::target AccessTarget = accessTarget>
-  typename std::enable_if<(AccessTarget == access::target::constant_buffer),
-                          constant_ptr<DataT>>::type
-  get_pointer() const {
-    return constant_ptr<DataT>(this->__get_impl()->Data);
-  }
-  /* Available only when: accessTarget == access::target::local */
-  template <typename DataT = typename detail::remove_AS<dataT>::type,
-            access::target AccessTarget = accessTarget>
-  typename std::enable_if<(AccessTarget == access::target::local),
-                          local_ptr<DataT>>::type
-  get_pointer() const {
-    return local_ptr<DataT>(this->__get_impl()->Data);
-  }
+    template <int CurDims = SubDims,
+              typename = enable_if_t<IsAccessReadOnly && CurDims == 1>>
+    DataT operator[](size_t Index) const {
+      MIDs[Dims - SubDims] = Index;
+      return MAccessor[MIDs];
+    }
+  };
 };
 
 } // namespace detail
 
-//
-// Actual definition of sycl::accessor class.
-//
-template <typename dataT, int dimensions, access::mode accessMode,
-          access::target accessTarget = access::target::global_buffer,
-          access::placeholder isPlaceholder = access::placeholder::false_t>
-class accessor
-    : public detail::accessor_pointer<
-          typename detail::DeviceValueType<dataT, accessTarget>::type,
-          dimensions, accessMode, accessTarget, isPlaceholder> {
-  using _ValueType =
-      typename detail::DeviceValueType<dataT, accessTarget>::type;
-  using _ImplT = detail::accessor_impl<_ValueType, dimensions, accessMode,
-                                       accessTarget, isPlaceholder>;
-
-  // Make sure Impl field is the first in the class, so that it is
-  // safe to reinterpret a pointer to accessor as a pointer to the
-  // implementation.
-#ifdef __SYCL_DEVICE_ONLY__
-  _ImplT impl;
-#else
-  std::shared_ptr<_ImplT> impl;
-  char padding[sizeof(_ImplT) - sizeof(std::shared_ptr<_ImplT>)];
-#endif
-
-#ifdef __SYCL_DEVICE_ONLY__
-  void __init(_ValueType *Ptr, range<dimensions> AccessRange,
-              range<dimensions> MemRange, id<dimensions> Offset) {
-    impl.Data = Ptr;
-    impl.AccessRange = AccessRange;
-    impl.MemRange = MemRange;
-    impl.Offset = Offset;
-  }
-#endif
-
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::target AccessTarget, access::placeholder IsPlaceholder>
+class accessor :
 #ifndef __SYCL_DEVICE_ONLY__
-  detail::buffer_impl<buffer_allocator> *getBufImpl() const {
-    return impl->m_Buf;
+    public detail::AccessorBaseHost,
+#endif
+    public detail::accessor_common<DataT, Dimensions, AccessMode, AccessTarget,
+                                   IsPlaceholder> {
+
+  static_assert((AccessTarget == access::target::global_buffer ||
+                 AccessTarget == access::target::constant_buffer ||
+                 AccessTarget == access::target::host_buffer),
+                "Expected buffer type");
+
+  template <bool B, class T = void>
+  using enable_if_t = typename std::enable_if<B, T>::type;
+
+  using AccessorCommonT = detail::accessor_common<DataT, Dimensions, AccessMode,
+                                                  AccessTarget, IsPlaceholder>;
+
+  constexpr static int AdjustedDim = Dimensions == 0 ? 1 : Dimensions;
+
+  using AccessorCommonT::AS;
+  using AccessorCommonT::IsAccessAnyWrite;
+  using AccessorCommonT::IsAccessReadOnly;
+  using AccessorCommonT::IsConstantBuf;
+  using AccessorCommonT::IsGlobalBuf;
+  using AccessorCommonT::IsHostBuf;
+  using AccessorCommonT::IsPlaceH;
+  template <int Dims>
+  using AccessorSubscript =
+      typename AccessorCommonT::template AccessorSubscript<Dims>;
+
+  using RefType = typename detail::PtrValueType<DataT, AS>::type &;
+  using PtrType = typename detail::PtrValueType<DataT, AS>::type *;
+
+  template <int Dims = Dimensions> size_t getLinearIndex(id<Dims> Id) const {
+
+#ifdef __SYCL_DEVICE_ONLY__
+    // Pointer is already adjusted for 1D case.
+    if (Dimensions == 1)
+      return Id[0];
+#endif // __SYCL_DEVICE_ONLY__
+
+    size_t Result = 0;
+    for (int I = 0; I < Dims; ++I)
+      Result = Result * getMemoryRange()[I] + getOffset()[I] + Id[I];
+    return Result;
   }
-#endif
 
-  range<dimensions> getAccessRange() const {
 #ifdef __SYCL_DEVICE_ONLY__
-    return impl.AccessRange;
+
+  id<AdjustedDim> &getOffset() { return impl.Offset; }
+  range<AdjustedDim> &getAccessRange() { return impl.AccessRange; }
+  range<AdjustedDim> &getMemoryRange() { return impl.MemRange; }
+
+  const id<AdjustedDim> &getOffset() const { return impl.Offset; }
+  const range<AdjustedDim> &getAccessRange() const { return impl.AccessRange; }
+  const range<AdjustedDim> &getMemoryRange() const { return impl.MemRange; }
+
+  detail::AccessorImplDevice<AdjustedDim> impl;
+
+  PtrType MData;
+
+  void __init(PtrType Ptr, range<AdjustedDim> AccessRange,
+              range<AdjustedDim> MemRange, id<AdjustedDim> Offset) {
+    MData = Ptr;
+    for (int I = 0; I < AdjustedDim; ++I) {
+      getOffset()[I] = Offset[I];
+      getAccessRange()[I] = AccessRange[I];
+      getMemoryRange()[I] = MemRange[I];
+    }
+    // In case of 1D buffer, adjust pointer during initialization rather
+    // then each time in operator[] or get_pointer functions.
+    if (1 == AdjustedDim)
+      MData += Offset[0];
+  }
+
+  PtrType getQualifiedPtr() const { return MData; }
 #else
-    return impl->AccessRange;
-#endif
-  };
 
-  range<dimensions> getMemRange() const {
-#ifdef __SYCL_DEVICE_ONLY__
-    return impl.MemRange;
-#else
-    return impl->MemRange;
-#endif
-  };
+  using AccessorBaseHost::getAccessRange;
+  using AccessorBaseHost::getOffset;
+  using AccessorBaseHost::getMemoryRange;
 
-  id<dimensions> getOffset() const {
-#ifdef __SYCL_DEVICE_ONLY__
-    return impl.Offset;
-#else
-    return impl->Offset;
-#endif
-  };
+  char padding[sizeof(detail::AccessorImplDevice<AdjustedDim>) +
+               sizeof(PtrType) - sizeof(detail::AccessorBaseHost)];
 
-  size_t getByteSize() const {
-#ifdef __SYCL_DEVICE_ONLY__
-    return impl.ByteSize;
-#else
-    return impl->ByteSize;
-#endif
-  };
+  PtrType getQualifiedPtr() const {
+    return reinterpret_cast<PtrType>(AccessorBaseHost::getPtr());
+  }
 
-  template <typename KernelType, int Dimensions, typename RangeType,
-            typename KernelArgType, bool SingleTask>
-  friend class cl::sycl::simple_scheduler::ExecuteKernelCommand;
-
-  template <int AccessDimensions, typename KernelType>
-  friend uint32_t cl::sycl::simple_scheduler::passGlobalAccessorAsArg(
-      uint32_t I, int LambdaOffset, cl_kernel ClKernel,
-      const KernelType &HostKernel);
-
-  template <int AccessDimensions, typename KernelType>
-  friend uint32_t cl::sycl::simple_scheduler::passLocalAccessorAsArg(
-      uint32_t I, int LambdaOffset, cl_kernel ClKernel,
-      const KernelType &HostKernel);
-
-  template <class Obj>
-  friend decltype(Obj::impl) detail::getSyclObjImpl(const Obj &SyclObject);
+#endif // __SYCL_DEVICE_ONLY__
 
 public:
-  using value_type = dataT;
-  using reference = dataT &;
-  using const_reference = const dataT &;
+  using value_type = DataT;
+  using reference = DataT &;
+  using const_reference = const DataT &;
 
-  // buffer accessor ctor #1
-  //   accessor(buffer<dataT, 1> &);
-  //
-  // Available only when:
-  //   ((isPlaceholder == access::placeholder::false_t &&
-  //     accessTarget == access::target::host_buffer) ||
-  //    (isPlaceholder == access::placeholder::true_t  &&
-  //     (accessTarget == access::target::global_buffer||
-  //      accessTarget == access::target::constant_buffer))) &&
-  //   dimensions == 0
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-           (((IsPlaceholder == access::placeholder::false_t &&
-              AccessTarget == access::target::host_buffer) ||
-             (IsPlaceholder == access::placeholder::true_t  &&
-               (AccessTarget == access::target::global_buffer ||
-                AccessTarget == access::target::constant_buffer))) &&
-            Dimensions == 0),
-           buffer<DataT, 1>>::type &bufferRef)
+  template <int Dims = Dimensions>
+  accessor(enable_if_t<((!IsPlaceH && IsHostBuf) ||
+                        (IsPlaceH && (IsGlobalBuf || IsConstantBuf))) &&
+                       Dims == 0, buffer<DataT, 1> > &BufferRef)
 #ifdef __SYCL_DEVICE_ONLY__
-      : impl(reinterpret_cast<_ValueType *>(
-                 detail::getSyclObjImpl(bufferRef)->BufPtr)) {
+      : impl(id<AdjustedDim>(), BufferRef.get_range(), BufferRef.MemRange) {
 #else
-      : impl(std::make_shared<_ImplT>(
-                 reinterpret_cast<_ValueType *>(
-                     detail::getSyclObjImpl(bufferRef)->BufPtr))) {
+      : AccessorBaseHost(
+            /*Offset=*/{0, 0, 0},
+            detail::convertToArrayOfN<3, 1>(BufferRef.get_range()),
+            detail::convertToArrayOfN<3, 1>(BufferRef.MemRange), AccessMode,
+            detail::getSyclObjImpl(BufferRef).get(), AdjustedDim,
+            sizeof(DataT)) {
+    detail::EventImplPtr Event =
+        detail::Scheduler::getInstance().addHostAccessor(
+            AccessorBaseHost::impl.get());
+    Event->wait(Event);
 #endif
-    auto BufImpl = detail::getSyclObjImpl(bufferRef);
-    if (AccessTarget == access::target::host_buffer) {
-      simple_scheduler::Scheduler::getInstance()
-          .copyBack<AccessMode, AccessTarget>(*BufImpl);
-    }
-    if (BufImpl->OpenCLInterop && !BufImpl->isValidAccessToMem(accessMode)) {
-      throw cl::sycl::runtime_error(
-          "Access mode is incompatible with opencl memory object of the "
-          "interoperability buffer");
-    }
   }
 
-  // buffer accessor ctor #2:
-  //   accessor(buffer<dataT, 1> &, handler &);
-  //
-  // Available only when:
-  //   isPlaceholder == access::placeholder::false_t &&
-  //   (accessTarget == access::target::global_buffer ||
-  //    accessTarget == access::target::constant_buffer) &&
-  //   dimensions == 0
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-           (IsPlaceholder == access::placeholder::false_t &&
-            (AccessTarget == access::target::global_buffer ||
-             AccessTarget == access::target::constant_buffer) &&
-             Dimensions == 0),
-           buffer<DataT, 1>>::type &bufferRef,
-           handler &commandGroupHandlerRef)
+  template <int Dims = Dimensions>
+  accessor(
+      buffer<DataT, 1> &BufferRef,
+      enable_if_t<(!IsPlaceH && (IsGlobalBuf || IsConstantBuf)) && Dims == 0,
+                   handler> &CommandGroupHandler)
 #ifdef __SYCL_DEVICE_ONLY__
-      // Even though this ctor can not be used in device code, some
-      // dummy implementation is still needed.
-      // Pass nullptr as a pointer to mem and use buffers from the ctor
-      // arguments to avoid the need in adding utility functions for
-      // dummy/default initialization of range fields.
-      : impl(nullptr) {}
-#else // !__SYCL_DEVICE_ONLY__
-      : impl(std::make_shared<_ImplT>(
-                 reinterpret_cast<_ValueType *>(
-                     detail::getSyclObjImpl(bufferRef)->BufPtr))) {
-    auto BufImpl = detail::getSyclObjImpl(bufferRef);
-    if (BufImpl->OpenCLInterop && !BufImpl->isValidAccessToMem(accessMode)) {
-      throw cl::sycl::runtime_error(
-          "Access mode is incompatible with opencl memory object of the "
-          "interoperability buffer");
-    }
-    commandGroupHandlerRef.AddBufDep<AccessMode, AccessTarget>(*BufImpl);
-    impl->m_Buf = BufImpl.get();
+      : impl(id<AdjustedDim>(), BufferRef.get_range(), BufferRef.MemRange) {
   }
-#endif // !__SYCL_DEVICE_ONLY__
-
-  // buffer accessor ctor #3:
-  //   accessor(buffer &);
-  //
-  // Available only when:
-  //   ((isPlaceholder == access::placeholder::false_t &&
-  //     accessTarget == access::target::host_buffer) ||
-  //    (isPlaceholder == access::placeholder::true_t &&
-  //     (accessTarget == access::target::global_buffer ||
-  //      accessTarget == access::target::constant_buffer))) &&
-  //   dimensions > 0)
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-           (((IsPlaceholder == access::placeholder::false_t &&
-              AccessTarget == access::target::host_buffer) ||
-             (IsPlaceholder == access::placeholder::true_t &&
-              (AccessTarget == access::target::global_buffer ||
-               AccessTarget == access::target::constant_buffer))) &&
-            Dimensions > 0),
-           buffer<DataT, Dimensions>>::type &bufferRef)
-#ifdef __SYCL_DEVICE_ONLY__
-      : impl(reinterpret_cast<_ValueType *>(
-                 detail::getSyclObjImpl(bufferRef)->BufPtr),
-             bufferRef.MemRange, bufferRef.MemRange) {
 #else
-      : impl(std::make_shared<_ImplT>(
-                 reinterpret_cast<_ValueType *>(
-                     detail::getSyclObjImpl(bufferRef)->BufPtr),
-                 bufferRef.MemRange, bufferRef.MemRange)) {
-#endif
-    auto BufImpl = detail::getSyclObjImpl(bufferRef);
-    if (AccessTarget == access::target::host_buffer) {
-      simple_scheduler::Scheduler::getInstance()
-          .copyBack<AccessMode, AccessTarget>(*BufImpl);
-    }
-    if (BufImpl->OpenCLInterop && !BufImpl->isValidAccessToMem(accessMode)) {
-      throw cl::sycl::runtime_error(
-          "Access mode is incompatible with opencl memory object of the "
-          "interoperability buffer");
-    }
-  }
-
-  // buffer ctor #4:
-  //   accessor(buffer &, handler &);
-  //
-  // Available only when:
-  //   isPlaceholder == access::placeholder::false_t &&
-  //   (accessTarget == access::target::global_buffer ||
-  //    accessTarget == access::target::constant_buffer) &&
-  //   dimensions > 0
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-               (IsPlaceholder == access::placeholder::false_t &&
-                (AccessTarget == access::target::global_buffer ||
-                 AccessTarget == access::target::constant_buffer) &&
-                Dimensions > 0),
-               buffer<DataT, Dimensions>>::type &bufferRef,
-           handler &commandGroupHandlerRef)
-#ifdef __SYCL_DEVICE_ONLY__
-      // Even though this ctor can not be used in device code, some
-      // dummy implementation is still needed.
-      // Pass nullptr as a pointer to mem and use buffers from the ctor
-      // arguments to avoid the need in adding utility functions for
-      // dummy/default initialization of range fields.
-      : impl(nullptr, bufferRef.MemRange, bufferRef.MemRange) {}
-#else
-      : impl(std::make_shared<_ImplT>(
-                 reinterpret_cast<_ValueType *>(
-                     detail::getSyclObjImpl(bufferRef)->BufPtr),
-                 bufferRef.MemRange, bufferRef.MemRange)) {
-    auto BufImpl = detail::getSyclObjImpl(bufferRef);
-    if (BufImpl->OpenCLInterop && !BufImpl->isValidAccessToMem(accessMode)) {
-      throw cl::sycl::runtime_error(
-          "Access mode is incompatible with opencl memory object of the "
-          "interoperability buffer");
-    }
-    commandGroupHandlerRef.AddBufDep<AccessMode, AccessTarget>(*BufImpl);
-    impl->m_Buf = BufImpl.get();
+      : AccessorBaseHost(
+            /*Offset=*/{0, 0, 0},
+            detail::convertToArrayOfN<3, 1>(BufferRef.get_range()),
+            detail::convertToArrayOfN<3, 1>(BufferRef.MemRange), AccessMode,
+            detail::getSyclObjImpl(BufferRef).get(), Dimensions,
+            sizeof(DataT)) {
+    CommandGroupHandler.associateWithHandler(*this);
   }
 #endif
 
-  // accessor ctor #5:
-  //   accessor(buffer &, range Range, id Offset = {});
-  //
-  // Available only when:
-  //   (isPlaceholder == access::placeholder::false_t &&
-  //    accessTarget == access::target::host_buffer) ||
-  //   (isPlaceholder == access::placeholder::true_t &&
-  //    (accessTarget == access::target::global_buffer ||
-  //     accessTarget == access::target::constant_buffer) &&
-  //    dimensions > 0)
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-               ((IsPlaceholder == access::placeholder::false_t &&
-                 AccessTarget == access::target::host_buffer) ||
-                (IsPlaceholder == access::placeholder::true_t &&
-                 (AccessTarget == access::target::global_buffer ||
-                  AccessTarget == access::target::constant_buffer) &&
-                 Dimensions > 0)),
-               buffer<DataT, Dimensions>>::type &bufferRef,
-           range<Dimensions> Range, id<Dimensions> Offset = {})
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<((!IsPlaceH && IsHostBuf) ||
+                              (IsPlaceH && (IsGlobalBuf || IsConstantBuf))) &&
+                             (Dims > 0)>>
+  accessor(buffer<DataT, Dimensions> &BufferRef)
 #ifdef __SYCL_DEVICE_ONLY__
-      // Even though this ctor can not be used in device code, some
-      // dummy implementation is still needed.
-      // Pass nullptr as a pointer to mem and use buffers from the ctor
-      // arguments to avoid the need in adding utility functions for
-      // dummy/default initialization of range<Dimensions> and
-      // id<Dimension> fields.
-      : impl(nullptr, Range, bufferRef.MemRange, Offset) {}
-#else   // !__SYCL_DEVICE_ONLY__
-      : impl(std::make_shared<_ImplT>(
-                 reinterpret_cast<_ValueType *>(
-                     detail::getSyclObjImpl(bufferRef)->BufPtr),
-                 Range, bufferRef.MemRange, Offset)) {
-    auto BufImpl = detail::getSyclObjImpl(bufferRef);
-    if (AccessTarget == access::target::host_buffer) {
-      simple_scheduler::Scheduler::getInstance()
-          .copyBack<AccessMode, AccessTarget>(*BufImpl);
-    }
-    if (BufImpl->OpenCLInterop && !BufImpl->isValidAccessToMem(accessMode)) {
-      throw cl::sycl::runtime_error(
-          "Access mode is incompatible with opencl memory object of the "
-          "interoperability buffer");
-    }
+      : impl(id<Dimensions>(), BufferRef.get_range(), BufferRef.MemRange) {
   }
-#endif // !__SYCL_DEVICE_ONLY__
-
-  // buffer ctor #6:
-  //   accessor(buffer &, handler &, range Range, id Offset);
-  //
-  // Available only when:
-  //   isPlaceholder == access::placeholder::false_t &&
-  //   (accessTarget == access::target::global_buffer ||
-  //    accessTarget == access::target::constant_buffer) &&
-  //   dimensions > 0
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-               (IsPlaceholder == access::placeholder::false_t &&
-                (AccessTarget == access::target::global_buffer ||
-                 AccessTarget == access::target::constant_buffer) &&
-                Dimensions > 0),
-               buffer<DataT, Dimensions>>::type &bufferRef,
-           handler &commandGroupHandlerRef, range<Dimensions> Range,
-           id<Dimensions> Offset = {})
-#ifdef __SYCL_DEVICE_ONLY__
-      // Even though this ctor can not be used in device code, some
-      // dummy implementation is still needed.
-      // Pass nullptr as a pointer to mem and use buffers from the ctor
-      // arguments to avoid the need in adding utility functions for
-      // dummy/default initialization of range<Dimensions> and
-      // id<Dimension> fields.
-      : impl(nullptr, Range, bufferRef.MemRange, Offset) {}
-#else   // !__SYCL_DEVICE_ONLY__
-      : impl(std::make_shared<_ImplT>(
-                 reinterpret_cast<_ValueType *>(
-                     detail::getSyclObjImpl(bufferRef)->BufPtr),
-                 Range, bufferRef.MemRange, Offset)) {
-    auto BufImpl = detail::getSyclObjImpl(bufferRef);
-    if (BufImpl->OpenCLInterop && !BufImpl->isValidAccessToMem(accessMode)) {
-      throw cl::sycl::runtime_error(
-          "Access mode is incompatible with opencl memory object of the "
-          "interoperability buffer");
-    }
-    commandGroupHandlerRef.AddBufDep<AccessMode, AccessTarget>(*BufImpl);
-    impl->m_Buf = BufImpl.get();
-  }
-#endif // !__SYCL_DEVICE_ONLY__
-
-  // local accessor ctor #1
-  //   accessor(handler &);
-  // Available only when:
-  //   AccessTarget == access::target::local && Dimensions == 0
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<
-               (AccessTarget == access::target::local &&
-                Dimensions == 0), handler>::type &commandGroupHandlerRef)
-#ifdef __SYCL_DEVICE_ONLY__
-      : impl(commandGroupHandlerRef) {}
 #else
-      : impl(std::make_shared<_ImplT>(commandGroupHandlerRef)) {}
+      : AccessorBaseHost(
+            /*Offset=*/{0, 0, 0},
+            detail::convertToArrayOfN<3, 1>(BufferRef.get_range()),
+            detail::convertToArrayOfN<3, 1>(BufferRef.MemRange), AccessMode,
+            detail::getSyclObjImpl(BufferRef).get(), Dimensions,
+            sizeof(DataT)) {
+    detail::EventImplPtr Event =
+        detail::Scheduler::getInstance().addHostAccessor(
+            AccessorBaseHost::impl.get());
+    Event->wait(Event);
+  }
 #endif
 
-  // local accessor ctor #2
-  //   accessor(range allocationSize, handler &);
-  // Available only when:
-  //   AccessTarget == access::target::local && Dimensions > 0
-  template <typename DataT = dataT, int Dimensions = dimensions,
-            access::mode AccessMode = accessMode,
-            access::target AccessTarget = accessTarget,
-            access::placeholder IsPlaceholder = isPlaceholder>
-  accessor(typename std::enable_if<(AccessTarget == access::target::local &&
-                                    Dimensions > 0),
-                                   range<Dimensions>>::type allocationSize,
-           handler &commandGroupHandlerRef)
+  template <int Dims = Dimensions,
+            typename = enable_if_t<
+                (!IsPlaceH && (IsGlobalBuf || IsConstantBuf)) && (Dims > 0)>>
+  accessor(buffer<DataT, Dimensions> &BufferRef, handler &CommandGroupHandler)
 #ifdef __SYCL_DEVICE_ONLY__
-      : impl(allocationSize, commandGroupHandlerRef) {}
+      : impl(id<AdjustedDim>(), BufferRef.get_range(), BufferRef.MemRange) {
+  }
 #else
-      : impl(std::make_shared<_ImplT>(allocationSize,
-                                      commandGroupHandlerRef)) {}
+      : AccessorBaseHost(
+            /*Offset=*/{0, 0, 0},
+            detail::convertToArrayOfN<3, 1>(BufferRef.get_range()),
+            detail::convertToArrayOfN<3, 1>(BufferRef.MemRange), AccessMode,
+            detail::getSyclObjImpl(BufferRef).get(), Dimensions,
+            sizeof(DataT)) {
+    CommandGroupHandler.associateWithHandler(*this);
+  }
 #endif
 
-  accessor(const accessor &rhs) = default;
-  accessor(accessor &&rhs) = default;
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<((!IsPlaceH && IsHostBuf) ||
+                              (IsPlaceH && (IsGlobalBuf || IsConstantBuf))) &&
+                             (Dims > 0)>>
+  accessor(buffer<DataT, Dimensions> &BufferRef, range<Dimensions> AccessRange,
+           id<Dimensions> AccessOffset = {})
+#ifdef __SYCL_DEVICE_ONLY__
+      : impl(AccessOffset, AccessRange, BufferRef.MemRange) {
+  }
+#else
+      : AccessorBaseHost(detail::convertToArrayOfN<3, 0>(AccessOffset),
+                           detail::convertToArrayOfN<3, 1>(AccessRange),
+                           detail::convertToArrayOfN<3, 1>(BufferRef.MemRange),
+                           AccessMode, detail::getSyclObjImpl(BufferRef).get(),
+                           Dimensions, sizeof(DataT)) {
+    detail::EventImplPtr Event =
+        detail::Scheduler::getInstance().addHostAccessor(
+            AccessorBaseHost::impl.get());
+    Event->wait(Event);
+  }
+#endif
 
-  accessor &operator=(const accessor &rhs) = default;
-  accessor &operator=(accessor &&rhs) = default;
+  template <int Dims = Dimensions,
+            typename = enable_if_t<
+                (!IsPlaceH && (IsGlobalBuf || IsConstantBuf)) && (Dims > 0)>>
+  accessor(buffer<DataT, Dimensions> &BufferRef, handler &CommandGroupHandler,
+           range<Dimensions> AccessRange, id<Dimensions> AccessOffset = {})
+#ifdef __SYCL_DEVICE_ONLY__
+      : impl(AccessOffset, AccessRange, BufferRef.MemRange) {
+  }
+#else
+      : AccessorBaseHost(detail::convertToArrayOfN<3, 0>(AccessOffset),
+                           detail::convertToArrayOfN<3, 1>(AccessRange),
+                           detail::convertToArrayOfN<3, 1>(BufferRef.MemRange),
+                           AccessMode, detail::getSyclObjImpl(BufferRef).get(),
+                           Dimensions, sizeof(DataT)) {
+    CommandGroupHandler.associateWithHandler(*this);
+  }
+#endif
 
-  ~accessor() = default;
+  constexpr bool is_placeholder() const { return IsPlaceH; }
 
-  bool operator==(const accessor &rhs) const { return impl == rhs.impl; }
-  bool operator!=(const accessor &rhs) const { return !(*this == rhs); }
+  size_t get_size() const { return getMemoryRange().size() * sizeof(DataT); }
+
+  size_t get_count() const { return getMemoryRange().size(); }
+
+  template <int Dims = Dimensions, typename = enable_if_t<(Dims > 0)>>
+  range<Dimensions> get_range() const {
+    return detail::convertToArrayOfN<Dimensions, 1>(getAccessRange());
+  }
+
+  template <int Dims = Dimensions, typename = enable_if_t<(Dims > 0)>>
+  id<Dimensions> get_offset() const {
+    return detail::convertToArrayOfN<Dimensions, 0>(getOffset());
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessAnyWrite && Dims == 0>>
+  operator RefType() const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>());
+    return *(getQualifiedPtr() + LinearIndex);
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessAnyWrite && (Dims > 0)>>
+  RefType operator[](id<Dimensions> Index) const {
+    const size_t LinearIndex = getLinearIndex(Index);
+    return getQualifiedPtr()[LinearIndex];
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessAnyWrite && Dims == 1>>
+  RefType operator[](size_t Index) const {
+    const size_t LinearIndex = getLinearIndex(id<Dimensions>(Index));
+    return getQualifiedPtr()[LinearIndex];
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessReadOnly && Dims == 0>>
+  operator DataT() const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>());
+    return *(getQualifiedPtr() + LinearIndex);
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessReadOnly && (Dims > 0)>>
+  DataT operator[](id<Dimensions> Index) const {
+    const size_t LinearIndex = getLinearIndex(Index);
+    return getQualifiedPtr()[LinearIndex];
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessReadOnly && Dims == 1>>
+  DataT operator[](size_t Index) const {
+    const size_t LinearIndex = getLinearIndex(id<Dimensions>(Index));
+    return getQualifiedPtr()[LinearIndex];
+  }
+
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<AccessMode == access::mode::atomic && Dims == 0>>
+  operator atomic<DataT, AS>() const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>());
+    return atomic<DataT, AS>(
+        multi_ptr<DataT, AS>(getQualifiedPtr() + LinearIndex));
+  }
+
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<AccessMode == access::mode::atomic && (Dims > 0)>>
+  atomic<DataT, AS> operator[](id<Dimensions> Index) const {
+    const size_t LinearIndex = getLinearIndex(Index);
+    return atomic<DataT, AS>(
+        multi_ptr<DataT, AS>(getQualifiedPtr() + LinearIndex));
+  }
+
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<AccessMode == access::mode::atomic && Dims == 1>>
+  atomic<DataT, AS> operator[](size_t Index) const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>(Index));
+    return atomic<DataT, AS>(
+        multi_ptr<DataT, AS>(getQualifiedPtr() + LinearIndex));
+  }
+
+  template <int Dims = Dimensions, typename = enable_if_t<(Dims > 1)>>
+  typename AccessorCommonT::template AccessorSubscript<Dims - 1>
+  operator[](size_t Index) const {
+    return AccessorSubscript<Dims - 1>(*this, Index);
+  }
+
+  template <
+      access::target AccessTarget_ = AccessTarget,
+      typename = enable_if_t<AccessTarget_ == access::target::host_buffer>>
+  DataT *get_pointer() const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>());
+    return getQualifiedPtr() + LinearIndex;
+  }
+
+  template <
+      access::target AccessTarget_ = AccessTarget,
+      typename = enable_if_t<AccessTarget_ == access::target::global_buffer>>
+  global_ptr<DataT> get_pointer() const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>());
+    return global_ptr<DataT>(getQualifiedPtr() + LinearIndex);
+  }
+
+  template <
+      access::target AccessTarget_ = AccessTarget,
+      typename = enable_if_t<AccessTarget_ == access::target::constant_buffer>>
+  constant_ptr<DataT> get_pointer() const {
+    const size_t LinearIndex = getLinearIndex(id<AdjustedDim>());
+    return constant_ptr<DataT>(getQualifiedPtr() + LinearIndex);
+  }
+
+  bool operator==(const accessor &Rhs) const { return impl == Rhs.impl; }
+  bool operator!=(const accessor &Rhs) const { return !(*this == Rhs); }
 };
+
+// Local accessor
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::placeholder IsPlaceholder>
+class accessor<DataT, Dimensions, AccessMode, access::target::local,
+               IsPlaceholder> :
+#ifndef __SYCL_DEVICE_ONLY__
+    public detail::LocalAccessorBaseHost,
+#endif
+    public detail::accessor_common<DataT, Dimensions, AccessMode,
+                                   access::target::local, IsPlaceholder> {
+
+  constexpr static int AdjustedDim = Dimensions == 0 ? 1 : Dimensions;
+
+  using AccessorCommonT =
+      detail::accessor_common<DataT, Dimensions, AccessMode,
+                              access::target::local, IsPlaceholder>;
+
+  using AccessorCommonT::AS;
+  using AccessorCommonT::IsAccessAnyWrite;
+  template <int Dims>
+  using AccessorSubscript =
+      typename AccessorCommonT::template AccessorSubscript<Dims>;
+
+  using RefType = typename detail::PtrValueType<DataT, AS>::type &;
+  using PtrType = typename detail::PtrValueType<DataT, AS>::type *;
+
+  template <bool B, class T = void>
+  using enable_if_t = typename std::enable_if<B, T>::type;
+
+#ifdef __SYCL_DEVICE_ONLY__
+  detail::LocalAccessorBaseDevice<AdjustedDim> impl;
+
+  sycl::range<AdjustedDim> &getSize() { return impl.MemRange; }
+  const sycl::range<AdjustedDim> &getSize() const { return impl.MemRange; }
+
+  void __init(PtrType Ptr, range<AdjustedDim> AccessRange,
+              range<AdjustedDim> MemRange, id<AdjustedDim> Offset) {
+    MData = Ptr;
+    for (int I = 0; I < AdjustedDim; ++I)
+      getSize()[I] = AccessRange[I];
+  }
+
+  PtrType getQualifiedPtr() const { return MData; }
+
+  PtrType MData;
+
+#else
+
+  char padding[sizeof(detail::LocalAccessorBaseDevice<AdjustedDim>) +
+               sizeof(PtrType) - sizeof(detail::LocalAccessorBaseHost)];
+  using detail::LocalAccessorBaseHost::getSize;
+
+  PtrType getQualifiedPtr() const {
+    return reinterpret_cast<PtrType>(LocalAccessorBaseHost::getPtr());
+  }
+
+#endif // __SYCL_DEVICE_ONLY__
+
+  // Method which calculates linear offset for the ID using Range and Offset.
+  template <int Dims = AdjustedDim> size_t getLinearIndex(id<Dims> Id) const {
+    size_t Result = 0;
+    for (int I = 0; I < Dims; ++I)
+      Result = Result * getSize()[I] + Id[I];
+    return Result;
+  }
+
+public:
+  using value_type = DataT;
+  using reference = DataT &;
+  using const_reference = const DataT &;
+
+  template <int Dims = Dimensions, typename = enable_if_t<Dims == 0>>
+  accessor(handler &CommandGroupHandler)
+#ifdef __SYCL_DEVICE_ONLY__
+      : impl(range<AdjustedDim>{1}) {
+  }
+#else
+      : LocalAccessorBaseHost(range<3>{1, 1, 1}, AdjustedDim, sizeof(DataT)) {
+  }
+#endif
+
+  template <int Dims = Dimensions, typename = enable_if_t<(Dims > 0)>>
+  accessor(range<Dimensions> AllocationSize, handler &CommandGroupHandler)
+#ifdef __SYCL_DEVICE_ONLY__
+      : impl(AllocationSize) {
+  }
+#else
+      : LocalAccessorBaseHost(detail::convertToArrayOfN<3, 1>(AllocationSize),
+                              AdjustedDim, sizeof(DataT)) {
+  }
+#endif
+
+  size_t get_size() const { return getSize().size() * sizeof(DataT); }
+
+  size_t get_count() const { return getSize().size(); }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessAnyWrite && Dims == 0>>
+  operator RefType() const {
+    return *getQualifiedPtr();
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessAnyWrite && (Dims > 0)>>
+  RefType operator[](id<Dimensions> Index) const {
+    const size_t LinearIndex = getLinearIndex(Index);
+    return getQualifiedPtr()[LinearIndex];
+  }
+
+  template <int Dims = Dimensions,
+            typename = enable_if_t<IsAccessAnyWrite && Dims == 1>>
+  RefType operator[](size_t Index) const {
+    return getQualifiedPtr()[Index];
+  }
+
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<AccessMode == access::mode::atomic && Dims == 0>>
+  operator atomic<DataT, AS>() const {
+    return atomic<DataT, AS>(multi_ptr<DataT, AS>(getQualifiedPtr()));
+  }
+
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<AccessMode == access::mode::atomic && (Dims > 0)>>
+  atomic<DataT, AS> operator[](id<Dimensions> Index) const {
+    const size_t LinearIndex = getLinearIndex(Index);
+    return atomic<DataT, AS>(
+        multi_ptr<DataT, AS>(getQualifiedPtr() + LinearIndex));
+  }
+
+  template <
+      int Dims = Dimensions,
+      typename = enable_if_t<AccessMode == access::mode::atomic && Dims == 1>>
+  atomic<DataT, AS> operator[](size_t Index) const {
+    return atomic<DataT, AS>(multi_ptr<DataT, AS>(getQualifiedPtr() + Index));
+  }
+
+  template <int Dims = Dimensions, typename = enable_if_t<(Dims > 1)>>
+  typename AccessorCommonT::template AccessorSubscript<Dims - 1>
+  operator[](size_t Index) const {
+    return AccessorSubscript<Dims - 1>(*this, Index);
+  }
+
+  local_ptr<DataT> get_pointer() const {
+    return local_ptr<DataT>(getQualifiedPtr());
+  }
+
+  bool operator==(const accessor &Rhs) const { return impl == Rhs.impl; }
+  bool operator!=(const accessor &Rhs) const { return !(*this == Rhs); }
+};
+
+// Image accessor
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::target AccessTarget, access::placeholder IsPlaceholder>
+class image_accessor {
+  static_assert(AccessTarget == access::target::image ||
+                    AccessTarget == access::target::host_image ||
+                    AccessTarget == access::target::image_array,
+                "Expected image type");
+  // TODO: Check if placeholder is applicable here.
+public:
+  using value_type = DataT;
+  using reference = DataT &;
+  using const_reference = const DataT &;
+
+  /* Available only when: accessTarget == access::target::host_image */
+  // template <typename AllocatorT>
+  // accessor(image<dimensions, AllocatorT> &imageRef);
+  /* Available only when: accessTarget == access::target::image */
+  // template <typename AllocatorT>
+  // accessor(image<dimensions, AllocatorT> &imageRef,
+  // handler &commandGroupHandlerRef);
+
+  /* Available only when: accessTarget == access::target::image_array &&
+        dimensions < 3 */
+  // template <typename AllocatorT>
+  // accessor(image<dimensions + 1, AllocatorT> &imageRef,
+  // handler &commandGroupHandlerRef);
+
+  /* TODO -- common interface members -- */
+  // size_t get_size() const;
+
+  // size_t get_count() const;
+
+  /* Available only when: (accessTarget == access::target::image || accessTarget
+     == access::target::host_image) && accessMode == access::mode::read */
+  // template <typename coordT> dataT read(const coordT &coords) const;
+
+  /* Available only when: (accessTarget == access::target::image || accessTarget
+     == access::target::host_image) && accessMode == access::mode::read */
+  // template <typename coordT>
+  // dataT read(const coordT &coords, const sampler &smpl) const;
+
+  /* Available only when: (accessTarget == access::target::image || accessTarget
+      == access::target::host_image) && accessMode == access::mode::write ||
+      accessMode == access::mode::discard_write */
+  // template <typename coordT>
+  // void write(const coordT &coords, const dataT &color) const;
+
+  /* Available only when: accessTarget == access::target::image_array &&
+     dimensions < 3 */
+  //__image_array_slice__ operator[](size_t index) const;
+};
+
+// Image accessors
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::placeholder IsPlaceholder>
+class accessor<DataT, Dimensions, AccessMode, access::target::image,
+               IsPlaceholder>
+    : public image_accessor<DataT, Dimensions, AccessMode,
+                            access::target::image, IsPlaceholder> {};
+
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::placeholder IsPlaceholder>
+class accessor<DataT, Dimensions, AccessMode, access::target::host_image,
+               IsPlaceholder>
+    : public image_accessor<DataT, Dimensions, AccessMode,
+                            access::target::host_image, IsPlaceholder> {};
+
+template <typename DataT, int Dimensions, access::mode AccessMode,
+          access::placeholder IsPlaceholder>
+class accessor<DataT, Dimensions, AccessMode, access::target::image_array,
+               IsPlaceholder>
+    : public image_accessor<DataT, Dimensions, AccessMode,
+                            access::target::image_array, IsPlaceholder> {};
 
 } // namespace sycl
 } // namespace cl
 
-#undef SYCL_ACCESSOR_IMPL
-#undef SYCL_ACCESSOR_SUBCLASS
+
 
 namespace std {
-template <typename T, int Dimensions, cl::sycl::access::mode AccessMode,
+template <typename DataT, int Dimensions, cl::sycl::access::mode AccessMode,
           cl::sycl::access::target AccessTarget,
           cl::sycl::access::placeholder IsPlaceholder>
-struct hash<cl::sycl::accessor<T, Dimensions, AccessMode, AccessTarget,
+struct hash<cl::sycl::accessor<DataT, Dimensions, AccessMode, AccessTarget,
                                IsPlaceholder>> {
-  using AccType = cl::sycl::accessor<
-      T, Dimensions, AccessMode, AccessTarget, IsPlaceholder>;
-  using ImplType = cl::sycl::detail::accessor_impl<
-      T, Dimensions, AccessMode, AccessTarget, IsPlaceholder>;
+  using AccType = cl::sycl::accessor<DataT, Dimensions, AccessMode,
+                                     AccessTarget, IsPlaceholder>;
 
   size_t operator()(const AccType &A) const {
 #ifdef __SYCL_DEVICE_ONLY__
     // Hash is not supported on DEVICE. Just return 0 here.
     return 0;
 #else
-    std::shared_ptr<ImplType> ImplPtr = cl::sycl::detail::getSyclObjImpl(A);
-    return hash<std::shared_ptr<ImplType>>()(ImplPtr);
+    // getSyclObjImpl() here returns a pointer to either AccessorImplHost
+    // or LocalAccessorImplHost depending on the AccessTarget.
+    auto AccImplPtr = cl::sycl::detail::getSyclObjImpl(A);
+    return hash<decltype(AccImplPtr)>()(AccImplPtr);
 #endif
   }
 };
+
 } // namespace std
-#endif // SCHEDULER_20
