+++ sycl/include/CL/sycl/buffer.hpp
@@ -7,11 +7,7 @@
 //===----------------------------------------------------------------------===//
 
 #pragma once
-#ifdef SCHEDULER_20
-  #include <CL/sycl/detail/buffer_impl2.hpp>
-#else
-  #include <CL/sycl/detail/buffer_impl.hpp>
-#endif // SCHEDULER_20
+#include <CL/sycl/detail/buffer_impl.hpp>
 #include <CL/sycl/detail/common.hpp>
 #include <CL/sycl/exception.hpp>
 #include <CL/sycl/stl.hpp>
