+++ sycl/include/CL/sycl/detail/queue_impl.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include <CL/sycl/context.hpp>
-#include <CL/sycl/detail/scheduler/scheduler.h>
+#include <CL/sycl/detail/scheduler/scheduler.hpp>
 #include <CL/sycl/device.hpp>
 #include <CL/sycl/event.hpp>
 #include <CL/sycl/handler.hpp>
